+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The CHANGELOG is available on [Github](https://github.com/luc-tielen/souffle-has
 - Improved dead code elimination
 - Optimization passes:
   - HoistConstraints (faster searches by narrowing search-space as early as possible)
+- CLI: Allow emitting initial and transformed RA IR
 
 ### Changed